Rename `CheckoutViewController` to `CheckoutWebViewController`.
